Some refactoring.

+++ src/codetests.cpp
@@ -24,7 +24,7 @@ struct Target
    };
 
    Type type;
-   size_t id;
+   uint32_t id;
 };
 
 struct Value
